fix: rename price interface
